pyproject.toml
[project]
name = "waterq_auto_sync"
requires-python = ">=3.10,<3.15"
version = "0.1.0"
dependencies = [
    "dagster==1.12.14",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "waterq_auto_sync"
registry_modules = [
    "waterq_auto_sync.components.*",
]


